Move Data_Points_Registry to a /data sub-directory and rename to Endpoint_Registry.
